+++ Source/World/World.h
@@ -53,7 +53,7 @@ private:
 
 	long long int overworldDayCycle = 240000;
 
-	long long int overworldTime = 240000 * 10.2;
+	long long int overworldTime = (long long int) (240000.0 * 10.2);
 
 };
